[smh_app-62] * Move make_qr_code() to libs.qrcode module. * Revise function API so just one option: encode_base64=True — only one encoding possible, only provide PNG. * Add a test for this option, getting back raw PNG data
