+++ apps/org/utils.py
@@ -1,6 +1,3 @@
-import base64
-import io
-import qrcode
 import random
 import string
 
@@ -23,22 +20,3 @@ def set_unique_slug(instance, based_on_field='name'):
         slug = '{}-{}'.format(slug, ''.join(
             random.choices(string.ascii_letters + string.digits, k=20)))
     instance.slug = slug
-
-
-def make_qr_code(text, format='PNG', encoding='base64'):
-    """
-    Convert the given text to a QR-code. If encoding=='base64', return base64;
-    otherwise, return raw image file data as bytes.
-    """
-    qr = qrcode.QRCode(version=None)
-    qr.add_data(text)
-    qr.make(fit=True)  # is a Pillow Image
-    image = qr.make_image(fill_color="black", back_color="white")
-    f = io.BytesIO()
-    image.save(f, format=format)
-    f.seek(0)
-    b = f.read()
-    if encoding == 'base64':
-        return base64.b64encode(b).decode('utf-8')
-    else:
-        return b
